server/podbaby/requirements.txt
beautifulsoup4==4.5.1
decorator==4.0.10
Django==1.10
django-appconf==1.0.2
django-extensions==1.7.3
django-imagekit==3.3
djangorestframework==3.4.6
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
pilkit==1.1.13
Pillow==3.3.1
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
pyPodcastParser==2.0.0
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
lxml==3.6.4
